Refactor Mesh UV storage, expand Material, and cleanup
Mesh texture coordinates now use XRecord storage for DWG/DXF compatibility, replacing the old list. Material gains new properties and improved DxfCodeValue mapping. DwgWriter and related IO classes are refactored for clarity. Entity.Material now ensures registration in the document. XRecord and its Entry class are enhanced with better cloning and documentation. Tests are updated for the new APIs and round-trip correctness. General code cleanup and improved event handling included.
